85/fixes review
