+++ lib/pages/detail/polish_capital_grapho.dart
@@ -1,0 +1,69 @@
+import 'package:flutter/material.dart';
+import 'package:syncfusion_flutter_gauges/gauges.dart';
+
+class CustomRadialGauge extends StatelessWidget {
+  final double percentage;
+
+  CustomRadialGauge({required this.percentage});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        SizedBox(
+          height: 100.0,
+          width: 100.0,
+          child: SfRadialGauge(
+            axes: <RadialAxis>[
+              RadialAxis(
+                minimum: 0,
+                maximum: 100,
+                showLabels: false,
+                showTicks: false,
+                axisLineStyle: AxisLineStyle(
+                  thickness: 0.15,
+                  cornerStyle: CornerStyle.bothCurve,
+                  color: Color(0xFFF5DEDD),
+                  thicknessUnit: GaugeSizeUnit.factor,
+                ),
+                pointers: <GaugePointer>[
+                  RangePointer(
+                    value: percentage,
+                    cornerStyle: CornerStyle.bothCurve,
+                    width: 0.15,
+                    sizeUnit: GaugeSizeUnit.factor,
+                    color: Color(0xFFE1203E),
+                  ),
+                ],
+                annotations: <GaugeAnnotation>[
+                  GaugeAnnotation(
+                    positionFactor: 0.1,
+                    angle: 90,
+                    widget: Text(
+                      '${percentage.toInt()}%',
+                      style: TextStyle(
+                        fontSize: 24.0,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ],
+          ),
+        ),
+        Padding(
+          padding: const EdgeInsets.only(top: 0.0), 
+          child: Text(
+            'Polski kapitał',
+            style: TextStyle(
+              fontSize: 16.0,
+              fontWeight: FontWeight.w600,
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
